+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/RpcMethod.java
@@ -110,7 +110,6 @@ public enum RpcMethod {
   ETH_GET_UNCLE_BY_BLOCK_NUMBER_AND_INDEX("eth_getUncleByBlockNumberAndIndex"),
   ETH_GET_UNCLE_COUNT_BY_BLOCK_HASH("eth_getUncleCountByBlockHash"),
   ETH_GET_UNCLE_COUNT_BY_BLOCK_NUMBER("eth_getUncleCountByBlockNumber"),
-  ETH_MINING("eth_mining"),
   ETH_NEW_BLOCK_FILTER("eth_newBlockFilter"),
   ETH_NEW_FILTER("eth_newFilter"),
   ETH_NEW_PENDING_TRANSACTION_FILTER("eth_newPendingTransactionFilter"),