+++ ethereum/blockcreation/src/main/java/org/hyperledger/besu/ethereum/blockcreation/NoopMiningCoordinator.java
@@ -14,7 +14,6 @@
  */
 package org.hyperledger.besu.ethereum.blockcreation;
 
-import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Wei;
 import org.hyperledger.besu.ethereum.core.Block;
 import org.hyperledger.besu.ethereum.core.BlockHeader;
@@ -68,11 +67,6 @@ public class NoopMiningCoordinator implements MiningCoordinator {
   @Override
   public Wei getMinPriorityFeePerGas() {
     return miningConfiguration.getMinPriorityFeePerGas();
-  }
-
-  @Override
-  public Optional<Address> getCoinbase() {
-    return miningConfiguration.getCoinbase();
   }
 
   @Override